Deprecate namespace_to_dict
